+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Acid, Fire, and Electricity now have a glowing effect
 * Fixed a bug where stacking any particle on top of dry Sand would let Water go through that particle to turn the Sand wet
 * Added the Steam particle
+* Water stops moving when too cold
+* Added the Glitch particle
 
 # 1.0.0 (2020-07-12)
 * First release